finish download image script (use aria2c)
